requirements.txt
web3==6.0.0
prometheus_client==0.22.1
python-dotenv==1.1.1
